+++ frontend/src/main/scala/pl/edu/agh/sius/services/PiService.scala
@@ -3,13 +3,13 @@ package pl.edu.agh.sius.services
 import scala.collection.mutable
 
 object PiService {
-  type Listener = (Int, Int) => Any
+  type Listener = (Double, Double) => Any
 
-  private var totalPoints: Int = 0
-  private var pointsInCircle: Int = 0
+  private var totalPoints: Double = 0
+  private var pointsInCircle: Double = 0
   private val listeners: mutable.Set[Listener] = mutable.Set.empty
 
-  def update(totalPointsVal: Int, pointsInCircleVal: Int) = {
+  def update(totalPointsVal: Double, pointsInCircleVal: Double) = {
     totalPoints = totalPointsVal
     pointsInCircle = pointsInCircleVal
     listeners.foreach(_.apply(totalPoints, pointsInCircle))
